% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmatizer.R
\name{lemmatize}
\alias{lemmatize}
\title{An R wrapper for Jagger's lemmatizer}
\usage{
lemmatize(input, model_path = NULL, keep = NULL, concat = TRUE)
}
\arguments{
\item{input}{an input.}

\item{model_path}{a path to the model.}

\item{keep}{a vector of POS(s) to keep. Default is \code{NULL}.}

\item{concat}{logical. If TRUE, the function returns a concatenated string. Default is \code{TRUE}.}
}
\value{
a vector (if \code{concat = TRUE}) or a list (if \code{concat = FALSE}).
}
\description{
An R wrapper for Jagger's lemmatizer
}
\examples{
 data(sentence_example)
 res_lemmatize <- lemmatize(sentence_example$text)
}
