% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lemmatize_cpp_vec}
\alias{lemmatize_cpp_vec}
\title{Lemmatize (a vector input)}
\usage{
lemmatize_cpp_vec(inputs, model_path, keep_vec, keep_all)
}
\description{
Lemmatize (a vector input)
}
\keyword{internal}
