% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{pos}
\alias{pos}
\title{An R wrapper for Jagger's POS tagger}
\usage{
pos(input, model_path = NULL, keep = NULL, format = c("list", "data.frame"))
}
\arguments{
\item{input}{an input.}

\item{model_path}{a path to the model.}

\item{keep}{a vector of POS(s) to keep. Default is \code{NULL}.}

\item{format}{a format of the output. Default is \code{list}.}
}
\value{
a list object.
}
\description{
An R wrapper for Jagger's POS tagger
}
