% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pos_simple_cpp_vec}
\alias{pos_simple_cpp_vec}
\title{POS tagging in C++ (only token and pos)}
\usage{
pos_simple_cpp_vec(inputs, model_path)
}
\description{
POS tagging in C++ (only token and pos)
}
\keyword{internal}
