% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentence_example}
\alias{sentence_example}
\title{An example sentence}
\format{
A data.frame with a single row and a single column:
\describe{
\item{text}{a sentence in Japanese}
}
}
\source{
Aozora Bunko: https://www.aozora.gr.jp/
}
\usage{
sentence_example
}
\description{
An example sentence
}
\keyword{datasets}
