% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tokenize_cpp_vec}
\alias{tokenize_cpp_vec}
\title{Tokenizer (a vector input)}
\usage{
tokenize_cpp_vec(inputs, model_path, keep_vec, keep_all)
}
\description{
Tokenizer (a vector input)
}
\keyword{internal}
