% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{tokenize_tbl}
\alias{tokenize_tbl}
\title{An R wrapper for Jagger's tokenizer (a tibble input)}
\usage{
tokenize_tbl(tbl, column, model_path = NULL, keep = NULL)
}
\arguments{
\item{tbl}{a tibble.}

\item{column}{a column name of the tibble to tokenize.}

\item{model_path}{a path to the model.}

\item{keep}{a vector of POS(s) to keep. Default is \code{NULL}.}
}
\value{
a tibble.
}
\description{
An R wrapper for Jagger's tokenizer (a tibble input)
}
